refactor(cursor): extract input-mapping + value coercers from cursorToolNormalization
cursorToolNormalization is entirely pure functions; the per-tool-kind input
mapping family (mapCursorToolInput + the map*/normalize* helpers) moves to
cursorToolInputMapping.ts, and the shared value coercers (stringValue/
numericValue/stringArray) to cursorToolValueCoercion.ts so both the input
mapping and the remaining success-formatting reuse them without a cycle.

Behavior-preserving (cursor suites — 410 tests — pass untouched).
cursorToolNormalization 542 -> 358 LOC; structural counters held at 0.
