+++ src/providers/cursor/runtime/cursorToolValueCoercion.ts
@@ -1,0 +1,27 @@
+/** Shared value coercers for normalizing Cursor CLI tool-call payloads. */
+
+export function stringValue(value: unknown): string {
+  if (typeof value === 'string') return value;
+  if (typeof value === 'number' || typeof value === 'boolean') return String(value);
+  return '';
+}
+
+export function numericValue(value: unknown): number | null {
+  if (typeof value === 'number' && Number.isFinite(value)) return value;
+  if (typeof value === 'string') {
+    const parsed = Number(value);
+    if (Number.isFinite(parsed)) return parsed;
+  }
+  return null;
+}
+
+export function stringArray(value: unknown): string[] {
+  if (!Array.isArray(value)) return [];
+  const out: string[] = [];
+  for (const entry of value) {
+    if (typeof entry === 'string' && entry.trim().length > 0) {
+      out.push(entry);
+    }
+  }
+  return out;
+}
